+++ docs/api_reference/document_search/documents.md
@@ -8,18 +8,18 @@
 
 ::: ragbits.document_search.documents.element.Element
 
-::: ragbits.document_search.documents.sources.Source
+::: ragbits.core.sources.Source
 
-::: ragbits.document_search.documents.sources.AzureBlobStorageSource
+::: ragbits.core.sources.AzureBlobStorageSource
 
-::: ragbits.document_search.documents.sources.GCSSource
+::: ragbits.core.sources.GCSSource
 
-::: ragbits.document_search.documents.sources.GitSource
+::: ragbits.core.sources.GitSource
 
-::: ragbits.document_search.documents.sources.HuggingFaceSource
+::: ragbits.core.sources.HuggingFaceSource
 
-::: ragbits.document_search.documents.sources.LocalFileSource
+::: ragbits.core.sources.LocalFileSource
 
-::: ragbits.document_search.documents.sources.S3Source
+::: ragbits.core.sources.S3Source
 
-::: ragbits.document_search.documents.sources.WebSource
+::: ragbits.core.sources.WebSource
