+++ docs/quickstart/quickstart2_rag.md
@@ -43,7 +43,7 @@ We first need to direct Ragbits to the location of the documents to load them. T
 
 ```python
 from pathlib import Path
-from ragbits.document_search.documents.sources import LocalFileSource
+from ragbits.core.sources import LocalFileSource
 
 # Path to the directory with markdown files to ingest
 documents_path = Path(__file__).parent / "pb-source/en"
