+++ examples/evaluation/document-search/advanced/config/pipeline/source/hf.yaml
@@ -1,4 +1,4 @@
-type: ragbits.document_search.documents.sources.hf:HuggingFaceSource
+type: ragbits.core.sources.hf:HuggingFaceSource
 config:
   path: "micpst/hf-docs"
   split: "train[:5]"
