+++ examples/evaluation/document-search/basic/evaluate.py
@@ -57,7 +57,7 @@ config = {
                 },
             },
             "source": {
-                "type": "ragbits.document_search.documents.sources:HuggingFaceSource",
+                "type": "ragbits.core.sources:HuggingFaceSource",
                 "config": {
                     "path": "micpst/hf-docs",
                     "split": "train[:5]",
