+++ packages/ragbits-core/src/ragbits/core/sources/gcs.py
@@ -4,13 +4,11 @@ from pathlib import Path
 from typing import ClassVar
 
 from ragbits.core.audit import trace, traceable
-from ragbits.document_search.documents.sources import Source
-from ragbits.document_search.documents.sources.base import get_local_storage_dir
+from ragbits.core.sources.base import Source, get_local_storage_dir
+from ragbits.core.utils.decorators import requires_dependencies
 
 with suppress(ImportError):
     from gcloud.aio.storage import Storage as StorageClient
-
-from ragbits.core.utils.decorators import requires_dependencies
 
 
 class GCSSource(Source):
