+++ packages/ragbits-core/src/ragbits/core/sources/git.py
@@ -8,10 +8,9 @@ with suppress(ImportError):
     import git
 
 from ragbits.core.audit import traceable
+from ragbits.core.sources.base import Source, get_local_storage_dir
 from ragbits.core.utils.decorators import requires_dependencies
 from ragbits.document_search.documents.exceptions import SourceNotFoundError
-from ragbits.document_search.documents.sources import Source
-from ragbits.document_search.documents.sources.base import get_local_storage_dir
 
 # Constants for URI parts
 _REPO_AND_FILE_PARTS = 2
