+++ packages/ragbits-core/src/ragbits/core/sources/hf.py
@@ -5,10 +5,9 @@ from pathlib import Path
 from typing import ClassVar
 
 from ragbits.core.audit import trace, traceable
+from ragbits.core.sources.base import Source, get_local_storage_dir
 from ragbits.core.utils.decorators import requires_dependencies
 from ragbits.document_search.documents.exceptions import SourceConnectionError, SourceNotFoundError
-from ragbits.document_search.documents.sources import Source
-from ragbits.document_search.documents.sources.base import get_local_storage_dir
 
 with suppress(ImportError):
     from datasets import load_dataset
