+++ packages/ragbits-core/src/ragbits/core/sources/s3.py
@@ -11,9 +11,8 @@ with suppress(ImportError):
     from botocore.client import BaseClient
     from botocore.exceptions import ClientError, NoCredentialsError, PartialCredentialsError
 
+from ragbits.core.sources.base import Source, get_local_storage_dir
 from ragbits.core.utils.decorators import requires_dependencies
-from ragbits.document_search.documents.sources import Source
-from ragbits.document_search.documents.sources.base import get_local_storage_dir
 
 
 class S3Source(Source):
