+++ packages/ragbits-core/src/ragbits/core/sources/web.py
@@ -8,10 +8,9 @@ from urllib.parse import urlparse
 with suppress(ImportError):
     import aiohttp
 
+from ragbits.core.sources.base import Source, get_local_storage_dir
 from ragbits.core.utils.decorators import requires_dependencies
 from ragbits.document_search.documents.exceptions import SourceNotFoundError, WebDownloadError
-from ragbits.document_search.documents.sources import Source
-from ragbits.document_search.documents.sources.base import get_local_storage_dir
 
 
 class WebSource(Source):
