--- packages/ragbits-document-search/src/ragbits/document_search/documents/sources/__init__.py
@@ -1,19 +1,0 @@
-from ragbits.document_search.documents.sources.base import Source  # noqa: I001
-from ragbits.document_search.documents.sources.azure import AzureBlobStorageSource
-from ragbits.document_search.documents.sources.gcs import GCSSource
-from ragbits.document_search.documents.sources.git import GitSource
-from ragbits.document_search.documents.sources.hf import HuggingFaceSource
-from ragbits.document_search.documents.sources.local import LocalFileSource
-from ragbits.document_search.documents.sources.s3 import S3Source
-from ragbits.document_search.documents.sources.web import WebSource
-
-__all__ = [
-    "AzureBlobStorageSource",
-    "GCSSource",
-    "GitSource",
-    "HuggingFaceSource",
-    "LocalFileSource",
-    "S3Source",
-    "Source",
-    "WebSource",
-]
