+++ packages/ragbits-document-search/tests/integration/test_sources.py
@@ -3,9 +3,9 @@ from pathlib import Path
 
 import pytest
 
+from ragbits.core.sources.base import LOCAL_STORAGE_DIR_ENV
+from ragbits.core.sources.hf import HuggingFaceSource
 from ragbits.document_search.documents.exceptions import SourceNotFoundError
-from ragbits.document_search.documents.sources import HuggingFaceSource
-from ragbits.document_search.documents.sources.base import LOCAL_STORAGE_DIR_ENV
 
 from ..helpers import env_vars_not_set
