+++ packages/ragbits-document-search/tests/unit/test_aws_source.py
@@ -2,7 +2,7 @@ from unittest.mock import patch
 
 from sympy.testing import pytest
 
-from ragbits.document_search.documents.sources.s3 import S3Source
+from ragbits.core.sources.s3 import S3Source
 
 
 def test_id():
@@ -29,7 +29,7 @@ async def test_from_uri_with_prefix():
         "https://s3.us-west-2.amazonaws.com/bucket/path/to/files*",
         "https://bucket.s3-us-west-2.amazonaws.com/path/to/files*",
     ]
-    with patch("ragbits.document_search.documents.sources.s3.S3Source.list_sources") as mock_list_sources:
+    with patch("ragbits.core.sources.s3.S3Source.list_sources") as mock_list_sources:
         for path in good_paths:
             await S3Source.from_uri(path)
             mock_list_sources.assert_called_with(bucket_name="bucket", prefix="path/to/files")
