+++ packages/ragbits-document-search/tests/unit/test_document_search.py
@@ -9,6 +9,8 @@ from unittest.mock import AsyncMock
 import pytest
 
 from ragbits.core.embeddings.noop import NoopEmbedder
+from ragbits.core.sources.gcs import GCSSource
+from ragbits.core.sources.local import LocalFileSource
 from ragbits.core.vector_stores.in_memory import InMemoryVectorStore
 from ragbits.document_search import DocumentSearch
 from ragbits.document_search._main import SearchConfig
@@ -18,13 +20,10 @@ from ragbits.document_search.documents.document import (
     DocumentType,
 )
 from ragbits.document_search.documents.element import TextElement
-from ragbits.document_search.documents.sources import GCSSource, LocalFileSource
 from ragbits.document_search.ingestion.parsers import DocumentParser
 from ragbits.document_search.ingestion.parsers.base import TextDocumentParser
 from ragbits.document_search.ingestion.parsers.router import DocumentParserRouter
-from ragbits.document_search.ingestion.strategies.batched import (
-    BatchedIngestStrategy,
-)
+from ragbits.document_search.ingestion.strategies.batched import BatchedIngestStrategy
 
 CONFIG = {
     "vector_store": {
@@ -452,10 +451,8 @@ async def test_document_search_ingest_from_huggingface_uri_basic():
             file.write("HuggingFace test content")
 
         with (
-            mock.patch("ragbits.document_search.documents.sources.hf.load_dataset", return_value=dataset),
-            mock.patch(
-                "ragbits.document_search.documents.sources.base.get_local_storage_dir", return_value=storage_dir
-            ),
+            mock.patch("ragbits.core.sources.hf.load_dataset", return_value=dataset),
+            mock.patch("ragbits.core.sources.base.get_local_storage_dir", return_value=storage_dir),
         ):
             document_search = DocumentSearch(
                 vector_store=vector_store,
