+++ packages/ragbits-document-search/tests/unit/test_documents.py
@@ -1,12 +1,12 @@
 import tempfile
 from pathlib import Path
 
+from ragbits.core.sources.local import LocalFileSource
 from ragbits.document_search.documents.document import (
     DocumentMeta,
     DocumentType,
     TextDocument,
 )
-from ragbits.document_search.documents.sources import LocalFileSource
 
 
 async def test_loading_local_file_source():
