+++ packages/ragbits-document-search/tests/unit/test_source_discriminator.py
@@ -4,8 +4,8 @@ from typing import Annotated
 import pydantic
 import pytest
 
-from ragbits.document_search.documents.sources import LocalFileSource, Source
-from ragbits.document_search.documents.sources.base import SourceDiscriminator
+from ragbits.core.sources.base import Source, SourceDiscriminator
+from ragbits.core.sources.local import LocalFileSource
 
 
 class ModelWithSource(pydantic.BaseModel):
